requirements.txt
flask
flask-restful
python-jose
requests
